Adding SLHC prims. (#140)
The dense pairwise distances API in cuML currently relies on CUTLASS and will soon be re-written to remove that dependency. Until then, the SLHC implementation in RAFT cannot depend on that API. In the meantime, I've abstracted that piece out to make it pluggable on the cuML side.

This PR depends on the changes in #139.

Authors:
  - Corey J. Nolet (@cjnolet)
  - Divye Gala (@divyegala)

Approvers:
  - Divye Gala (@divyegala)
  - Dante Gama Dessavre (@dantegd)

URL: https://github.com/rapidsai/raft/pull/140