fix logic bug in include_checker.py utility (#734)
Hi, I noticed an issue in the include style checker when I was trying to adapt it for use in another project.

As far as I can tell, the current logic results in no source files actually being checked! Fortunately, the script also passes after the proposed change, so the includes were already in good shape :)

I verified the fix here by changing a <...>-style import to a double-quote one in one of the source files and it was detected.

I don't know if this file originates in RAFT or if it is also in other RAPIDS projects as well? If the later, the fix will need to be propagated to other repositories.

Authors:
  - Gregory Lee (https://github.com/grlee77)

Approvers:
  - Corey J. Nolet (https://github.com/cjnolet)

URL: https://github.com/rapidsai/raft/pull/734