mdspan-ify permute
Add overloads of permute that take arrays as mdspan
(std::optional<mdspan<...>> if the array is an optional
output argument).

Improve documentation of the existing overload
and document the new overloads.  Add tests.
Make sure that the function compiles if users pass in
std::nullopt for either or both optional arguments.