Fixing the sample-without-replacement test failures (#1109)
The sample-without-replacement test fails on A100 when `largeWeight` parameter is enabled, observed in PR #1076. The failure is caused as the sample with largest weight might not always be the first element getting selected in the sampled array. To fix this issue, this PR proposes to search the sample with large weight in the first three elements of sampled array.

Authors:
  - Vinay Deshpande (https://github.com/vinaydes)

Approvers:
  - Dante Gama Dessavre (https://github.com/dantegd)

URL: https://github.com/rapidsai/raft/pull/1109