Make IVF-PQ build index in batches when necessary (#1056)
Before this patch, when the input data was not accessible directly from the device, the `build` and `extend` functions mapped it using the `cudaHostRegister`. Although this approach was rather fast, it could fail when the input data is too large to fit in the device memory.
This PR, changes the logic of `build` and `extend`, so that the data is loaded in batches when necessary. Moreover,  when the passed pointer represents the mapped file (e.g. using the system call `mmap` ), the size of the input may even be larger than the host memory.
The `build` does one pass through the input (to sample the training set), and the `extend` does at most two passes.

Authors:
  - Artem M. Chirkin (https://github.com/achirkin)
  - Tamas Bela Feher (https://github.com/tfeher)

Approvers:
  - Tamas Bela Feher (https://github.com/tfeher)

URL: https://github.com/rapidsai/raft/pull/1056