Add ability to specify select_k algorithm
This adds the ability to specify which select_k algorithm we are using in
`matrix::select_k` - with the goal of being able to run experiments to
automatically pick the best algorithm.

We currently have at least 3 different implementations of a top-k algorithm
inside of RAFT : radix select, warp sort - and the block select from faiss.
(there is also another version in CAGRA that isn't included here).  Each
of these algorithms is fastest on certain inputs, and slower than others
on other inputs.